+++ .editorconfig
@@ -49,6 +49,6 @@ indent_size = 4
 indent_style = space
 indent_size = 2
 
-[**/nginx{/*,.*}.conf]
+[**/ngin{x,x/*,x.*}.conf]
 indent_style = space
 indent_size = 2
