+++ .env
@@ -4,8 +4,12 @@ MEDEA_CONF=_dev/config.toml
 MEDEA_CONTROL__STATIC_SPECS_DIR=_dev/specs/
 
 COMPOSE_PROJECT_NAME=medea
+
 COMPOSE_IMAGE_NAME=instrumentisto/medea
 COMPOSE_IMAGE_VER=dev
 
-MEDEA_CONTROL_MOCK_IMAGE_NAME=instrumentisto/medea-control-api-mock
-MEDEA_CONTROL_MOCK_IMAGE_VER=dev
+COMPOSE_CONTROL_MOCK_IMAGE_NAME=instrumentisto/medea-control-api-mock
+COMPOSE_CONTROL_MOCK_IMAGE_VER=dev
+
+COMPOSE_WEBDRIVER_IMAGE_NAME=selenoid/chrome
+COMPOSE_WEBDRIVER_IMAGE_VER=latest
