+++ .rustfmt.toml
@@ -2,7 +2,7 @@
 # https://github.com/rust-lang-nursery/rustfmt/blob/master/Configurations.md
 
 max_width = 80
-format_strings = true
+format_strings = false
 imports_granularity = "Crate"
 normalize_comments = true
 wrap_comments = true
