+++ Cargo.toml
@@ -104,7 +104,7 @@ webdriver = "0.43"
 [dev-dependencies.tokio_1]
     package = "tokio"
     version = "1.1"
-    features = ["macros", "rt-multi-thread", "fs"]
+    features = ["fs", "macros", "rt-multi-thread"]
 
 [[test]]
 name = "e2e"
