+++ tests/e2e/control.rs
@@ -1,4 +1,4 @@
-//! HTTP client that interacts with Medea via Control API.
+//! HTTP client interacting with Medea via its Control API.
 
 use derive_more::{Display, Error, From};
 use medea_control_api_mock::{
@@ -8,7 +8,7 @@ use medea_control_api_mock::{
 
 use crate::conf;
 
-/// All errors which can happen while working with Control API.
+/// All errors which can happen while working with a Control API.
 #[derive(Debug, Display, Error, From)]
 pub enum Error {
     Reqwest(reqwest::Error),
@@ -16,16 +16,19 @@ pub enum Error {
 
 type Result<T> = std::result::Result<T, Error>;
 
-/// Client for the Control API.
+/// Client of a Control API.
 pub struct Client(reqwest::Client);
 
 impl Client {
-    /// Returns new [`ControlApi`] client.
+    /// Returns a new Control API [`Client`].
+    #[inline]
+    #[must_use]
     pub fn new() -> Self {
         Self(reqwest::Client::new())
     }
 
-    /// Creates provided [`Element`] in the provided `path`.
+    /// Creates the provided media [`Element`] in the provided `path` on a Medea
+    /// media server.
     pub async fn create(
         &self,
         path: &str,
@@ -41,20 +44,19 @@ impl Client {
             .await?)
     }
 
-    /// Deletes [`Element`] in the provided `path`.
-    #[allow(dead_code)]
+    /// Deletes a media [`Element`] identified by the provided `path`.
     pub async fn delete(&self, path: &str) -> Result<Response> {
         Ok(self.0.delete(&get_url(path)).send().await?.json().await?)
     }
 
-    /// Returns [`Element`] from the provided `path`.
-    #[allow(dead_code)]
+    /// Returns a media [`Element`] identified by the provided `path`.
     pub async fn get(&self, path: &str) -> Result<SingleGetResponse> {
         Ok(self.0.get(&get_url(path)).send().await?.json().await?)
     }
 }
 
-/// Returns URL to the Control API for the provided [`Element`] path.
+/// Returns URL of a media [`Element`] identified by the provided `path`.
+#[must_use]
 fn get_url(path: &str) -> String {
     format!("{}/{}", *conf::CONTROL_API_ADDR, path)
 }
