+++ tests/e2e/nginx.conf
@@ -1,25 +1,27 @@
-worker_processes  1;
+worker_processes    1;
 events {
-    worker_connections  1024;
+  worker_connections    1024;
 }
 
 http {
-    default_type  application/octet-stream;
-    sendfile        on;
-    keepalive_timeout  65;
+  default_type    application/octet-stream;
+  sendfile        on;
 
-    server {
-        listen       30000;
-        server_name  localhost;
+  keepalive_timeout    65;
 
-        location / {
-            root   /usr/share/nginx/html;
-            index  index.html index.htm;
-            types {
-                application/javascript js;
-                application/wasm wasm;
-                text/html html;
-            }
-        }
+  server {
+    listen         30000;
+    server_name    localhost;
+
+    location / {
+      root     /usr/share/nginx/html;
+      index    index.html index.htm;
+
+      types {
+        application/javascript   js;
+        application/wasm         wasm;
+        text/html                html;
+      }
     }
+  }
 }
