+++ tests/e2e/object/room.rs
@@ -1,43 +1,38 @@
-//! Implementation and definition for the object which represents `Room` JS
-//! object.
+//! [`Object`] representing a `Room` JS object.
 
 use crate::{
     browser::Statement,
     object::{connections_store::ConnectionStore, Object},
 };
 
-/// Representation of the `Room` JS object.
+/// Representation of a `Room` JS object.
 pub struct Room;
 
-/// Representation of the `MediaKind` JS enum.
+/// Representation of a `MediaKind` JS enum.
 pub enum MediaKind {
     Audio,
     Video,
 }
 
-/// Representation of the `MediaSourceKind` JS enum.
-#[allow(dead_code)]
+/// Representation of a `MediaSourceKind` JS enum.
 pub enum MediaSourceKind {
     Device,
     Display,
 }
 
 impl MediaSourceKind {
-    /// Converts this [`MediaSourceKind`] to the JS code for this enum variant.
+    /// Converts this [`MediaSourceKind`] to a JS code for this enum variant.
     fn as_js(&self) -> String {
         match self {
-            MediaSourceKind::Device => {
-                "window.rust.MediaSourceKind.DEVICE".to_string()
-            }
-            MediaSourceKind::Display => {
-                "window.rust.MediaSourceKind.DISPLAY".to_string()
-            }
+            MediaSourceKind::Device => "window.rust.MediaSourceKind.DEVICE",
+            MediaSourceKind::Display => "window.rust.MediaSourceKind.DISPLAY",
         }
+        .to_owned()
     }
 }
 
 impl Object<Room> {
-    /// Joins [`Room`] with a provided URI.
+    /// Joins a [`Room`] with the provided URI.
     pub async fn join(&self, uri: String) -> Result<(), super::Error> {
         self.execute(Statement::new(
             // language=JavaScript
@@ -47,17 +42,16 @@ impl Object<Room> {
                     await room.join(uri);
                 }
             "#,
-            vec![uri.into()],
+            [uri.into()],
         ))
         .await?;
-
         Ok(())
     }
 
-    /// Disabled media publishing for the provided [`MediaKind`] and
+    /// Disables media publishing for the provided [`MediaKind`] and
     /// [`MediaSourceKind`].
     ///
-    /// If provided [`None`] `source_kind` then media publishing will be
+    /// If the provided `source_kind` is [`None`], then media publishing will be
     /// disabled for all [`MediaSourceKind`]s.
     pub async fn disable_media_send(
         &self,
@@ -68,7 +62,7 @@ impl Object<Room> {
             .as_ref()
             .map_or_else(String::new, MediaSourceKind::as_js);
         let disable = match kind {
-            MediaKind::Audio => "room.disable_audio()".to_string(),
+            MediaKind::Audio => "room.disable_audio()".to_owned(),
             MediaKind::Video => {
                 format!("room.disable_video({})", media_source_kind)
             }
@@ -81,16 +75,15 @@ impl Object<Room> {
                         await {};
                     }}
                 "#,
-                disable
+                disable,
             ),
-            vec![],
+            [],
         ))
         .await?;
-
         Ok(())
     }
 
-    /// Returns [`ConnectionStore`] for this [`Room`].
+    /// Returns a [`ConnectionStore`] of this [`Room`].
     pub async fn connections_store(
         &self,
     ) -> Result<Object<ConnectionStore>, super::Error> {
@@ -100,21 +93,20 @@ impl Object<Room> {
                 async (room) => {
                     let store = {
                         connections: new Map(),
-                        subs: new Map(),
+                        subs: new Map()
                     };
                     room.on_new_connection((conn) => {
                         let id = conn.get_remote_member_id();
                         store.connections.set(id, conn);
                         let sub = store.subs.get(id);
-                        if (sub != undefined) {
+                        if (sub !== undefined) {
                             sub(conn);
                         }
                     });
-
                     return store;
                 }
             "#,
-            vec![],
+            [],
         ))
         .await
     }
