+++ tests/e2e/steps/connection.rs
@@ -2,30 +2,31 @@ use cucumber_rust::then;
 
 use crate::world::World;
 
-#[then(regex = "^(\\S*) receives connection with (\\S*)$")]
+#[then(regex = r"^(\S+) receives connection with (\S+)$")]
 async fn then_member_receives_connection(
     world: &mut World,
     id: String,
-    partner_id: String,
+    responder_id: String,
 ) {
     let member = world.get_member(&id).unwrap();
     member
         .connections()
-        .wait_for_connection(partner_id.clone())
+        .wait_for_connection(responder_id.clone())
         .await
         .unwrap();
+    assert!(true);
 }
 
-#[then(regex = "^(\\S*) doesn't receives connection with (\\S*)")]
-async fn then_member_doesnt_receives_connection(
+#[then(regex = r"^(\S+) doesn't receive connection with (\S+)$")]
+async fn then_member_doesnt_receive_connection(
     world: &mut World,
     id: String,
-    partner_id: String,
+    responder_id: String,
 ) {
     let member = world.get_member(&id).unwrap();
     assert!(member
         .connections()
-        .get(partner_id)
+        .get(responder_id)
         .await
         .unwrap()
         .is_none())
